+++ src/ja2/editor/editor-map-info-c.ts
@@ -116,17 +116,17 @@ function ExtractAndUpdateMapInfo(): void {
   temp = min(GetNumericStrictValueFromField(1), 255);
   if (temp != -1 && temp != gEditorLightColor.peRed) {
     fUpdateLight1 = TRUE;
-    gEditorLightColor.peRed = (UINT8)temp;
+    gEditorLightColor.peRed = temp;
   }
   temp = min(GetNumericStrictValueFromField(2), 255);
   if (temp != -1 && temp != gEditorLightColor.peGreen) {
     fUpdateLight1 = TRUE;
-    gEditorLightColor.peGreen = (UINT8)temp;
+    gEditorLightColor.peGreen = temp;
   }
   temp = min(GetNumericStrictValueFromField(3), 255);
   if (temp != -1 && temp != gEditorLightColor.peBlue) {
     fUpdateLight1 = TRUE;
-    gEditorLightColor.peBlue = (UINT8)temp;
+    gEditorLightColor.peBlue = temp;
   }
   if (fUpdateLight1) {
     gfEditorForceShadeTableRebuild = TRUE;
@@ -137,21 +137,21 @@ function ExtractAndUpdateMapInfo(): void {
   // extract radius
   temp = max(min(GetNumericStrictValueFromField(4), 8), 1);
   if (temp != -1)
-    gsLightRadius = (INT16)temp;
+    gsLightRadius = temp;
   temp = max(min(GetNumericStrictValueFromField(5), 15), 1);
   if (temp != -1 && temp != gusLightLevel) {
-    gusLightLevel = (UINT16)temp;
+    gusLightLevel = temp;
     gfRenderWorld = TRUE;
-    ubAmbientLightLevel = (UINT8)(EDITOR_LIGHT_MAX - gusLightLevel);
+    ubAmbientLightLevel = (EDITOR_LIGHT_MAX - gusLightLevel);
     LightSetBaseLevel(ubAmbientLightLevel);
     LightSpriteRenderAll();
   }
 
-  temp = (INT8)GetNumericStrictValueFromField(6);
+  temp = GetNumericStrictValueFromField(6);
   if (temp == -1)
     gMapInformation.ubRestrictedScrollID = 0;
   else
-    gMapInformation.ubRestrictedScrollID = (UINT8)temp;
+    gMapInformation.ubRestrictedScrollID = temp;
 
   // set up fields for exitgrid information
   Get16BitStringFromField(7, str);
@@ -159,15 +159,15 @@ function ExtractAndUpdateMapInfo(): void {
     str[0] -= 32; // uppercase it!
   if (str[0] >= 'A' && str[0] <= 'Z' && str[1] >= '0' && str[1] <= '9') {
     // only update, if coordinate is valid.
-    gExitGrid.ubGotoSectorY = (UINT8)(str[0] - 'A' + 1);
-    gExitGrid.ubGotoSectorX = (UINT8)(str[1] - '0');
+    gExitGrid.ubGotoSectorY = (str[0] - 'A' + 1);
+    gExitGrid.ubGotoSectorX = (str[1] - '0');
     if (str[2] >= '0' && str[2] <= '9')
-      gExitGrid.ubGotoSectorX = (UINT8)(gExitGrid.ubGotoSectorX * 10 + str[2] - '0');
-    gExitGrid.ubGotoSectorX = (UINT8)max(min(gExitGrid.ubGotoSectorX, 16), 1);
-    gExitGrid.ubGotoSectorY = (UINT8)max(min(gExitGrid.ubGotoSectorY, 16), 1);
+      gExitGrid.ubGotoSectorX = (gExitGrid.ubGotoSectorX * 10 + str[2] - '0');
+    gExitGrid.ubGotoSectorX = max(min(gExitGrid.ubGotoSectorX, 16), 1);
+    gExitGrid.ubGotoSectorY = max(min(gExitGrid.ubGotoSectorY, 16), 1);
   }
-  gExitGrid.ubGotoSectorZ = (UINT8)max(min(GetNumericStrictValueFromField(8), 3), 0);
-  gExitGrid.usGridNo = (UINT16)max(min(GetNumericStrictValueFromField(9), 25600), 0);
+  gExitGrid.ubGotoSectorZ = max(min(GetNumericStrictValueFromField(8), 3), 0);
+  gExitGrid.usGridNo = max(min(GetNumericStrictValueFromField(9), 25600), 0);
 
   UpdateMapInfoFields();
 }
